Django_LoveSync_Diary/code/DjangoPro1/requirements.txt
Django==4.2.7
uuid==1.30
psycopg2-binary==2.9.9
Pillow==10.1.0
django-filter==2.5.0
django-rest-framework==3.15.2
mysqlclient==2.2.7
Pillow==11.3.0
django_redis==6.0.0
redis==6.2.0